Refactor AlertMessage component; move to new directory structure and restore styles
