+++ client/src/components/AlertMessage/AlertMessage.css
@@ -1,0 +1,66 @@
+.alert {
+  display: flex;
+  align-items: center;
+  margin-top: 10px;
+  padding: 10px;
+  border-radius: 5px;
+  font-size: 14px;
+  animation: fadeSlide 0.4s ease-out;
+}
+
+.alert-success {
+  padding: 0.75rem;
+  background-color: #f0fdf4;
+  border-left: 3px solid #22c55e;
+  border-radius: 0.375rem;
+  color: #166534;
+  font-size: 0.875rem;
+  margin-top: 0.5rem;
+}
+
+.alert-error {
+  padding: 0.75rem;
+  background-color: #fef2f2;
+  border-left: 3px solid #dc2626;
+  border-radius: 0.375rem;
+  color: #991b1b;
+  font-size: 0.875rem;
+  margin-top: 0.5rem;
+}
+
+.alert-warning {
+  padding: 0.75rem;
+  background-color: #fffbeb;
+  border-left: 3px solid #f59e0b;
+  border-radius: 0.375rem;
+  color: #92400e;
+  font-size: 0.875rem;
+  margin-top: 0.5rem;
+}
+
+.alert-info {
+  padding: 0.75rem;
+  background-color: #eff6ff;
+  border-left: 3px solid #3b82f6;
+  border-radius: 0.375rem;
+  color: #1e40af;
+  font-size: 0.875rem;
+  margin-top: 0.5rem;
+}
+
+.icon {
+  margin-right: 8px;
+  font-size: 16px;
+}
+
+/* Fade and slide animation */
+@keyframes fadeSlide {
+  0% {
+    opacity: 0;
+    transform: translateY(-10px);
+  }
+  100% {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
