+++ client/src/components/SearchInput.jsx
@@ -1,7 +1,7 @@
 import { useState, useRef } from "react";
 import { useNavigate } from "react-router-dom";
 import { UseJobs } from "../context/JobsContext";
-import AlertMessage from "../components/AlertMessage";
+import AlertMessage from "./AlertMessage/AlertMessage";
 import { validateJobInput } from "../util/searchValidation";
 import "./SearchInput.css";
 import { cleanUpText } from "../util/cleanUpText";
